requirements.txt
# Minervini SEPA Agentic Trading Workflow - Dependencies

# Data Sources
yfinance>=0.2.36
nsetools>=1.0.11
bsedata>=0.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
ta>=0.11.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# UI
streamlit>=1.30.0
plotly>=5.18.0

# HTTP/API
requests>=2.31.0
aiohttp>=3.9.0

# Logging & Utilities
loguru>=0.7.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
ipython>=8.0.0
jupyter>=1.0.0
aisuite
